refacto inject file to typescript
