+++ openaev-front/src/actions/communications/communication-helper.d.ts
@@ -1,3 +1,6 @@
 import { type Communication } from '../../utils/api-types';
 
-export interface CommunicationHelper { getExerciseCommunications: (exerciseId: string) => Communication[] }
+export interface CommunicationHelper {
+  getExerciseCommunications: (exerciseId: string) => Communication[];
+  getInjectCommunications: (injectId: string) => Communication[];
+}
